Private types and functions

Documentation for ClimaOceanBiogeochemistry.jl's internal interface.

ClimaOceanBiogeochemistry

ClimaOceanBiogeochemistry.CAN вҖ” Method

Tracer sources and sinks for alkalinity (ALK)

ClimaOceanBiogeochemistry.CAN вҖ” Method

Tracer sources and sinks for dissolved inorganic carbon (DIC)

ClimaOceanBiogeochemistry.CAN вҖ” Method

Tracer sources and sinks for dissolved organic phosphorus (DOP)

ClimaOceanBiogeochemistry.CAN вҖ” Method

Tracer sources and sinks for dissolved iron (FeT)

ClimaOceanBiogeochemistry.CAN вҖ” Method

Tracer sources and sinks for dissolved inorganic nitrate (NOвӮғ)

ClimaOceanBiogeochemistry.CAN вҖ” Method

Tracer sources and sinks for Particulate Organic Phosphorus (POP).

ClimaOceanBiogeochemistry.CAN вҖ” Method

Tracer sources and sinks for dissolved inorganic phosphate (POвӮ„)

ClimaOceanBiogeochemistry.CarbonAlkalinityNutrients вҖ” Method

CarbonAlkalinityNutrients(  reference_density                             = 1024.5,
                            maximum_net_community_production_rate         = 1 / day,
                            phosphate_half_saturation                     = 1e-7 * reference_density,
                            nitrate_half_saturation                       = 1.6e-6 * reference_density,
                            iron_half_saturation                          = 1e-10 * reference_density,
                            incident_PAR                                  = 700.0,
                            PAR_half_saturation                           = 10.0,
                            PAR_attenuation_scale                         = 25.0,
                            PAR_percent                                   = 0.01,
                            fraction_of_particulate_export                = 0.33,
                            dissolved_organic_phosphorus_remin_timescale  = 1 / 30day,
                            option_of_particulate_remin                   = 1,
                            particulate_organic_phosphorus_remin_timescale= 0.03 / day, 
                            stoichoimetric_ratio_carbon_to_phosphate      = 106.0
                            stoichoimetric_ratio_nitrate_to_phosphate     = 16.0
                            stoichoimetric_ratio_phosphate_to_oxygen      = 170.0,
                            stoichoimetric_ratio_iron_to_phosphate        = 4.68e-4
                            stoichoimetric_ratio_carbon_to_nitrate        = 106 / 16
                            stoichoimetric_ratio_carbon_to_oxygen         = 106 / 170,
                            stoichoimetric_ratio_carbon_to_iron           = 106 / 1.e-3,
                            stoichoimetric_ratio_silicate_to_phosphate    = 15.0,
                            rain_ratio_inorganic_to_organic_carbon        = 1e-1,
                            martin_curve_exponent                         = 0.84, 
                            iron_scavenging_rate                          = 5e-4 / day,
                            ligand_concentration                          = 1e-9 * reference_density,
                            ligand_stability_coefficient                  = 1e8
                            particulate_organic_phosphorus_sinking_velocity    = -10.0 / day)

Return a seven-tracer biogeochemistry model for the interaction of carbon, alkalinity, and nutrients.

Keyword Arguments

Tracer names

DIC: Dissolved Inorganic Carbon

ALK: Alkalinity

POвӮ„: Phosphate (macronutrient)

NOвӮғ: Nitrate (macronutrient)

DOP: Dissolved Organic Phosphorus

POP: Particulate Organic Phosphorus

Fe: Dissolved Iron (micronutrient)

Biogeochemical functions

transitions for DIC, ALK, POвӮ„, NOвӮғ, DOP, POP and Fe

biogeochemical_drift_velocity for POP, modeling the sinking of detritus at a constant detritus_sinking_speed.

ClimaOceanBiogeochemistry.NutrientsPlanktonBacteriaDetritus вҖ” Method

NutrientsPlanktonBacteriaDetritus(; grid,
                                    maximum_plankton_growth_rate = 1/day,
                                    maximum_bacteria_growth_rate = 1/day
                                    maximum_grazing_rate         = 3/day
                                    bacteria_yield               = 0.2
                                    zooplankton_yield            = 0.3
                                    linear_remineralization_rate = 0.03/day,
                                    linear_mortality_rate        = 0.01/day,
                                    quadratic_mortality_rate     = 0.1/day,
                                    quadratic_mortality_rate_Z   = 1/day,
                                    nutrient_half_saturation     = 0.1,
                                    detritus_half_saturation     = 0.1,
                                    grazing_half_saturation      = 3.0,
                                    PAR_half_saturation          = 10.0,
                                    PAR_attenuation_scale        = 25.0,
                                    detritus_vertical_velocity   = -10/day)

Return a six-tracer biogeochemistry model for the interaction of nutrients (N), phytoplankton (P), zooplankton(Z), bacteria (B), dissolved detritus (D1), and particulate detritus (D2).

Keyword Arguments

grid (required): An Oceananigans' grid.

maximum_plankton_growth_rate: (sвҒ»В№) Growth rate of plankton P unlimited by the availability of nutrients and light. Default: 1/day.

maximum_bacteria_growth_rate: (sвҒ»В№) Growth rate of plankton B unlimited by the availability of nutrients and light. Default = 0.5/day.

maximum_grazing_rate: (sвҒ»В№) Maximum grazing rate of phytoplankton by zooplankton.

bacteria_yield: Determines fractional nutrient production by bacteria production relative to consumption of detritus such that $вҲӮ_t N / вҲӮ_t D = 1 - y$, where y = bacteria_yield. Default: 0.2.

linear_remineralization_rate: (sвҒ»В№) Remineralization rate constant of detritus 'D', assuming linear remineralization of 'D', while implicitly modeling bacteria 'B'. Default = 0.3/day.

linear_mortality_rate: (sвҒ»В№) Linear term of the mortality rate of both plankton and bacteria.

quadratic_mortality_rate: (sвҒ»В№) Quadratic term of the mortality rate of both plankton and bacteria.

nutrient_half_saturation: (mmol mвҒ»Ві) Half-saturation of nutrients for plankton production.

detritus_half_saturation: (mmol mвҒ»Ві) Half-saturation of nutrients for bacteria production. Default = 10.0 mmol mвҒ»Ві.

phytoplankton_half_saturation: (mmol mвҒ»Ві) Half-saturation of phytoplankton for zooplankton production.

zooplankton_assimilation: Fractional assimilation efficiency for zooplankton.

PAR_half_saturation: (W mвҒ»ВІ) Half-saturation of photosynthetically available radiation (PAR) for plankton production.

PAR_attenuation_scale: (m) Depth scale over which photosynthetically available radiation (PAR) attenuates exponentially.

detritus_sinking_speed: (m sвҒ»В№) Sinking velocity of particulate detritus.

Tracer names

N: nutrients

P: phytoplankton

Z: zooplankton

B: bacteria

D: detritus

Biogeochemical functions

transitions for N, P, Z, B, D

biogeochemical_drift_velocity for D2, modeling the sinking of detritus at a constant detritus_sinking_speed.

ClimaOceanBiogeochemistry.PAR вҖ” Method

PAR(surface_photosynthetically_active_ratiation, 
    photosynthetically_active_ratiation_attenuation_scale, 
    depth)

Calculate the photosynthetically active radiation (PAR) at a given depth due to attenuation.

ClimaOceanBiogeochemistry.dissolved_organic_phosphorus_remin вҖ” Method

dissolved_organic_phosphorus_remin(remineralization_rate, 
                                  dissolved_organic_phosphorus_concentration)

Calculate the remineralization of dissolved organic phosphorus.

ClimaOceanBiogeochemistry.iron_scavenging вҖ” Method

iron_scavenging(iron_scavenging_rate, 
                iron_concentration, 
                ligand_concentration, 
                ligand_stability_coefficient)

Calculate the scavenging loss of iron. Iron scavenging depends on free iron, which involves solving a quadratic equation in terms of ligand concentration and stability coefficient. Ligand-complexed iron is protected from being scavenged.

ClimaOceanBiogeochemistry.iron_sources вҖ” Method

Add surface input of iron. This sould be a boundary condition, but for now we just add a constant source.

ClimaOceanBiogeochemistry.net_community_production вҖ” Method

net_community_production(maximum_net_community_production_rate,
                         light_half_saturation, 
                         phosphate_half_saturation, 
                         nitrate_half_saturation, 
                         iron_half_saturation, 
                         photosynthetically_active_radiation, 
                         phosphate_concentration, 
                         nitrate_concentration, 
                         iron_concentration)

ClimaOceanBiogeochemistry.particulate_inorganic_carbon_remin вҖ” Method

Calculate remineralization of particulate inorganic carbon.

ClimaOceanBiogeochemistry.particulate_organic_phosphorus_remin вҖ” Method

Calculate remineralization of particulate organic phosphorus according to 1) rate decreases with depth (1/z+zвӮҖ), or 2) a first-order rate constant.

Oceananigans.Biogeochemistry.biogeochemical_drift_velocity вҖ” Method

Add a vertical sinking "drift velocity" for the particulate organic phosphate (POP) tracer.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.BбөҖбҙјбөҖ вҖ” Function

BбөҖбҙјбөҖ(Sбө–, Pбҙ®бөҖбҙјбөҖ)

Return total borate concentration in mol/kg-SW given practical salinity, Sбө–. References: UppstrГ¶m (1974), cited by Dickson et al. (2007, chapter 5, p 10) Millero (1982) cited in Millero (1995)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.CaбөҖбҙјбөҖ вҖ” Function

CaбөҖбҙјбөҖ(Sбө–, Pб¶ңбөғбөҖбҙјбөҖ)

Return calcium concentration in mol/kg-SW given practical salinity, Sбө–. References: Culkin and Cox (1966), Culkin (1967), Riley and Tongudai (1967)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FCбөҖCOвӮӮЛўбө’ЛЎ вҖ” Method

FCбөҖCOвӮӮЛўбө’ЛЎ(CбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the dissolved and hydrated COвӮӮ concentration in seawater given the total carbon concentration CбөҖ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FCбөҖCOвӮғВІвҒ» вҖ” Method

FCбөҖCOвӮғВІвҒ»(CбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the carbonate concentration in seawater given the total carbon concentration CбөҖ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FCбөҖHCOвӮғвҒ» вҖ” Method

FCбөҖHCOвӮғвҒ»(CбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the bicarbonate ion concentration in seawater given the total carbon concentration CбөҖ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FpCOвӮӮCOвӮӮЛўбө’ЛЎ вҖ” Method

FpCOвӮӮCOвӮӮЛўбө’ЛЎ(pCOвӮӮ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the dissolved and hydrated COвӮӮ concentration in seawater given the pCOвӮӮ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FpCOвӮӮCOвӮғВІвҒ» вҖ” Method

FpCOвӮӮCOвӮғВІвҒ»(pCOвӮӮ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the carbonate concentration in seawater given the pCOвӮӮ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FpCOвӮӮHCOвӮғвҒ» вҖ” Method

FpCOвӮӮHCOвӮғвҒ»(pCOвӮӮ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the bicarbonate ion concentration in seawater given the pCOвӮӮ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FЛўвҒұбө—вӮ–вӮҒ вҖ” Function

FЛўвҒұбө—вӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PЛўвҒұбө—вӮ–вӮҒ)

Return the first dissociation constant of silicic acid (H4SiO4) in seawater, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PЛўвҒұбө—вӮ–вӮҒ.

References: Yao and Millero (1995) cited by Millero (1995) pH scale : SWS (according to Dickson et al, 2007) Note : No pressure correction available Note : converted here from mol/kg-H2O to mol/kg-sw

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбҙҙВІЛўвӮ–вӮҒ вҖ” Function

FбҙҙВІЛўвӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбҙҙВІЛўвӮ–вӮҒ)

Return the dissociation constant of hydrogen sulfide in sea-water, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбҙҙВІЛўвӮ–вӮҒ.

References: Millero et al. (1988) (cited by Millero (1995) Millero (1995) for pressure correction pH scale : - SWS (according to Yao and Millero, 1995, p. 82: "refitted if necessary") - Total (according to Lewis and Wallace, 1998) Note : we stick to SWS here for the time being Note : the fits from Millero (1995) and Yao and Millero (1995) derive from Millero et al. (1988), with all the coefficients multiplied by -ln(10)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбҙҙВІбҙјвӮ–вӮҒ вҖ” Function

FбҙҙВІбҙјвӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбҙҙВІбҙјвӮ–вӮҒ)

Return dissociation constant of water in (mol/kg-SW)^2, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбҙҙВІбҙјвӮ–вӮҒ.

References: Millero (1995) for value at p_bar = 0 Millero (pers. comm. 1996) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбҙҙЛўбҙјвҒҙвӮ–вӮҒ вҖ” Function

FбҙҙЛўбҙјвҒҙвӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбҙҙЛўбҙјвҒҙвӮ–вӮҒ)

Return the dissociation constant of hydrogen sulfate (bisulfate) , given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбҙҙЛўбҙјвҒҙвӮ–вӮҒ.

References: Dickson (1990) вҖ“ also Handbook (2007) Millero (1995) for pressure correction pH scale : free Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fбҙҙб¶ бөҰвӮҒ вҖ” Function

Fбҙҙб¶ бөҰвӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, Pбҙҙб¶ бөҰвӮҒ)

Return the association constant of HF in (mol/kg-SW)^-1, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, Pбҙҙб¶ бөҰвӮҒ.

HF <-> HвҒә + FвҒ»

References: Dickson and Riley (1979) Millero (1995) for pressure correction pH scale : free Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fбҙҙб¶ вӮ–вӮҒ вҖ” Function

Fбҙҙб¶ вӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, Pбҙҙб¶ вӮ–вӮҒ)

Return the dissociation constant for hydrogen fluoride in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, Pбҙҙб¶ вӮ–вӮҒ.

HF <-> HвҒә + FвҒ»

References: Perez and Fraga (1987) Millero (1995) for pressure correction pH scale : Total (according to Handbook, 2007

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбҙәбҙҙвҒҙвӮ–вӮҒ вҖ” Function

FбҙәбҙҙвҒҙвӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбҙәбҙҙвҒҙвӮ–вӮҒ)

Return the dissociation constant of ammonium in sea-water [mol/kg-SW], given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбҙәбҙҙвҒҙвӮ–вӮҒ.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбҙҫбҙјвҒҙвӮ–вӮҒ вҖ” Function

FбҙҫбҙјвҒҙвӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбҙҫбҙјвҒҙвӮ–вӮҒ)

Return the first dissociation constant of phosphoric acid (H3PO4) in seawater, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбҙҫбҙјвҒҙвӮ–вӮҒ.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбҙҫбҙјвҒҙвӮ–вӮӮ вҖ” Function

FбҙҫбҙјвҒҙвӮ–вӮӮ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбҙҫбҙјвҒҙвӮ–вӮӮ)

Return the second dissociation constant of phosphoric acid (H3PO4) in seawater, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбҙҫбҙјвҒҙвӮ–вӮӮ.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбҙҫбҙјвҒҙвӮ–вӮғ вҖ” Function

FбҙҫбҙјвҒҙвӮ–вӮғ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбҙҫбҙјвҒҙвӮ–вӮғ)

Return the third dissociation constant of phosphoric acid (H3PO4) in seawater, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбҙҫбҙјвҒҙвӮ–вӮғ.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҖбҙјбөҖ вҖ” Function

FбөҖбҙјбөҖ(Sбө–, Pб¶ бөҖбҙјбөҖ)

Return total fluoride concentration in mol/kg-SW given practical salinity, Sбө–. References: Culkin (1965) (???)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөғКібөғбөҚбө’вҒҝвҒұбө—бөүвӮӣвӮҡ вҖ” Function

FбөғКібөғбөҚбө’вҒҝвҒұбө—бөүвӮӣвӮҡ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбөғКібөғбөҚбө’вҒҝвҒұбө—бөүвӮӣвӮҡ)

Return stoichiometric solubility product, О©, of aragonite in seawater, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбөғКібөғбөҚбө’вҒҝвҒұбө—бөүвӮӣвӮҡ.

References: Mucci (1983) Millero (1979) for pressure correction pH scale : N/A Units : (mol/kg-SW)^2

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҮвӮ–вӮҒ вҖ” Function

FбөҮвӮ–вӮҒ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, Pбҙ®вӮ–вӮҒ)

Return boric acid dissociation constant in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, Pбҙ®вӮ–вӮҒ.

References: Dickson (1990, eq. 23) вҖ“ also Handbook (2007, eq. 37) Millero (1979) pressure correction pH scale : total

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮҖ вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮҖ(Оҳбҙ·, Sбө–, PбөҲвҒұб¶ңвӮ–вӮҖ)

Return hydration constant of COвӮӮ in (mol/kg-SW)/atm given temperature in K, Оҳбҙ·, practical salinity, Sбө–, and coefficients, PбөҲвҒұб¶ңвӮ–вӮҖ.

COвӮӮ + HвӮӮO <-> HвӮӮCOвӮғ

References: Weiss (1979) pH scale : N/A Note : currently no pressure correction

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮҒбөЈвӮүвӮғ вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮҒбөЈвӮүвӮғ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбөҲвҒұб¶ңвӮ–вӮҒбөЈвӮүвӮғ)

Return the first dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбөҲвҒұб¶ңвӮ–вӮҒбөЈвӮүвӮғ.

HвӮӮCOвӮғ <-> HCOвӮғвҒ» + HвҒә

References: Roy et al. (1993) вҖ“ also Handbook (1994) Millero (1979) pressure correction pH scale : Total Valid range: T: 0-45 S: 5-45. Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮҒвӮ—вӮҖвӮҖ вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮҒвӮ—вӮҖвӮҖ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбөҲвҒұб¶ңвӮ–вӮҒвӮ—вӮҖвӮҖ)

Return the first dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбөҲвҒұб¶ңвӮ–вӮҒвӮ—вӮҖвӮҖ.

HвӮӮCOвӮғ <-> HCOвӮғвҒ» + HвҒә

References: Luecker et al. (2000) вҖ“ also Handbook (2007) Millero (1979) pressure correction pH scale: Total

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮҒвӮҳвӮүвӮ… вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮҒвӮҳвӮүвӮ…(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбөҲвҒұб¶ңвӮ–вӮҒвӮҳвӮүвӮ…)

Return the first dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбөҲвҒұб¶ңвӮ–вӮҒвӮҳвӮүвӮ….

HвӮӮCOвӮғ <-> HCOвӮғвҒ» + HвҒә

References: Millero (1995, eq 50 вҖ“ ln K1(COM)) Millero (1982) pressure correction pH scale: SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮӮбөЈвӮүвӮғ вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮӮбөЈвӮүвӮғ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбөҲвҒұб¶ңвӮ–вӮӮбөЈвӮүвӮғ)

Return the second dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбөҲвҒұб¶ңвӮ–вӮӮбөЈвӮүвӮғ.

HCOвӮғвҒ» <-> COвӮғВІвҒ» + HвҒә

References: Roy et al. (1993) вҖ“ also Handbook (1994) Millero (1979) pressure correction pH scale : Total Valid range: T: 0-45 S: 5-45. Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮӮвӮ—вӮҖвӮҖ вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮӮвӮ—вӮҖвӮҖ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбөҲвҒұб¶ңвӮ–вӮӮвӮ—вӮҖвӮҖ)

Return the second dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбөҲвҒұб¶ңвӮ–вӮӮвӮ—вӮҖвӮҖ.

HCOвӮғвҒ» <-> COвӮғВІвҒ» + HвҒә

References: Luecker et al. (2000) вҖ“ also Handbook (2007) Millero (1979) pressure correction pH scale: Total

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮӮвӮҳвӮүвӮ… вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮӮвӮҳвӮүвӮ…(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, PбөҲвҒұб¶ңвӮ–вӮӮвӮҳвӮүвӮ…)

Return the second dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, PбөҲвҒұб¶ңвӮ–вӮӮвӮҳвӮүвӮ….

HCOвӮғвҒ» <-> COвӮғВІвҒ» + HвҒә

References: Millero (1995, eq 51 вҖ“ ln K2(COM)) Millero (1979) pressure correction pH scale: SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮҡбөЈвӮ‘ вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮҡбөЈвӮ‘(Оҳбҙ·, Sбө–, PбөҲвҒұб¶ңвӮ–вӮҡбөЈвӮ‘)

Return fugacity prefactor needed for non-ideality of COвӮӮ in the ocean in (mol/kg-SW)/atm given temperature in K, Оҳбҙ·, practical salinity, Sбө–, and coefficients, PбөҲвҒұб¶ңвӮ–вӮҡбөЈвӮ‘.

References: Weiss (1974) Marine Chemistry pH scale : N/A Note : currently no pressure correction

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FбөҲвҒұб¶ңвӮ–вӮӣвӮ’вӮ—вӮҗ вҖ” Function

FбөҲвҒұб¶ңвӮ–вӮӣвӮ’вӮ—вӮҗ(Оҳбҙ·, Sбө–, PбөҲвҒұб¶ңвӮ–вӮӣвӮ’вӮ—вӮҗ)

Calculate f = k0(1-pH2O)*correction term for non-ideality in (mol/kg-SW)/atm given temperature 
in K, `Оҳбҙ·`, practical salinity, `Sбө–`, and coefficients, `PбөҲвҒұб¶ңвӮ–вӮӣвӮ’вӮ—вӮҗ`. Currently no pressure correction

References: Weiss & Price (1980, Mar. Chem., 8, 347-359 Eq 13 with table 6 values)
pH scale  : N/A
Note      : currently no pressure correction

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fб¶ңбөғЛЎб¶ңвҒұбө—бөүвӮӣвӮҡ вҖ” Function

Fб¶ңбөғЛЎб¶ңвҒұбө—бөүвӮӣвӮҡ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real, Pб¶ңбөғЛЎб¶ңвҒұбө—бөүвӮӣвӮҡ)

Return the stoichiometric solubility product of calcite, О©, in seawater, given temperature in K, Оҳбҙ·, practical salinity, Sбө–, applied pressure, О”pбөҰвӮҗбөЈ, and coefficients, Pб¶ңбөғЛЎб¶ңвҒұбө—бөүвӮӣвӮҡ

References: Mucci (1983) Millero (1995) for pressure correction pH scale : N/A Units : (mol/kg-SW)^2

ClimaOceanBiogeochemistry.CarbonSystemSolvers.HвҒәвӮӣoverHвҒәвӮғ вҖ” Function

HвҒәвӮӣoverHвҒәвӮғ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real)

Return the ratio HSWS/Hfree as a function of salinity, Sбө–.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.HвҒәвӮӣoverHвҒәвӮң вҖ” Function

HвҒәвӮӣoverHвҒәвӮң(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real)

Return the ratio HSWS/HTot as a function of salinity, Sбө–. Reference: Munhoven pH scale: all

ClimaOceanBiogeochemistry.CarbonSystemSolvers.HвҒәвӮңoverHвҒәвӮғ вҖ” Function

HвҒәвӮңoverHвҒәвӮғ(Оҳбҙ·::Real, Sбө–::Real, О”pбөҰвӮҗбөЈ::Real)

Return the ratio HTot/Hfree as a function of salinity, Sбө–. Reference: Munhoven pH scale: N/A

ClimaOceanBiogeochemistry.CarbonSystemSolvers.HвӮӮOЛўК· вҖ” Function

HвӮӮOЛўК·(Sбө–, PбҙҙВІвҒ°ЛўК·)

Return the mass of pure water in one kg of seawater of practical salinity, Sбө–. References: "libthdyct" вҖ“ derived by Munhoven (1997) from data by Millero (1982) "Handbook (2007)" вҖ“ Handbook (2007) pH scale: N/A

ClimaOceanBiogeochemistry.CarbonSystemSolvers.SOвӮ„бөҖбҙјбөҖ вҖ” Function

SOвӮ„бөҖбҙјбөҖ(Sбө–, PЛўбҙјвҒҙбөҖбҙјбөҖ)

Return total sulfate concentration in mol/kg-SW given practical salinity, Sбө–. References: Morris, A.W. and Riley, J.P. (1966) quoted in Handbook (2007)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.ОјвӮӣ вҖ” Function

ОјвӮӣ(Sбө–)

Return ionic strength in mol/kg-SW, for given practical salinity, Sбө–. References: "libthdyct" вҖ“ derived by Munhoven (1997) from data by Millero (1982) "Handbook (2007)" вҖ“ Handbook (2007) pH scale: N/A

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FABбөҖ вҖ” Method

function FABбөҖ(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FACбөҖ вҖ” Method

function FACбөҖ(CбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAFбөҖ вҖ” Method

function FAFбөҖ(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAHвӮӮO вҖ” Method

function FAHвӮӮO(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAHвӮӮSбөҖ вҖ” Method

function FAHвӮӮSбөҖ(HвӮӮSбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FANHвӮ„бөҖ вҖ” Method

function FANHвӮ„бөҖ(NHвӮ„бөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAPбөҖ вҖ” Method

function FAPбөҖ(PбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FASOвӮ„бөҖ вҖ” Method

function FASOвӮ„бөҖ(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FASiбөҖ вҖ” Method

function FASiбөҖ(SiбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAбөҖ вҖ” Method

function FAбөҖ(CбөҖ, AбөҖ, BбөҖ, PбөҖ, SiбөҖ,  SOвӮ„бөҖ, FбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

Evaluate the rational function form of the total alkalinity-pH equation

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FHвҒәбөўвӮҷбөў вҖ” Method

FHвҒәбөўвӮҷбөў(AбөҖ, CбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

Calculates the root for the 2nd order approximation of the CбөҖ-BбөҖ-Aб¶ң equation for HвҒә (reformulated as a cubic polynomial) around the local minimum, if it exists.

Returns * 1e-03 if Aб¶ң <= 0 * 1e-10 if Aб¶ң >= 2CбөҖ + BбөҖ * 1e-07 if 0 < Aб¶ң < 2CбөҖ + BбөҖ and the 2nd order approximation does not have a solution

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FboundsAбөҖвӮҷвӮ•вӮӮвӮ’ вҖ” Method

 FboundsAбөҖвӮҷвӮ•вӮӮвӮ’(
    CбөҖ, PбөҖ, SiбөҖ, NHвӮ„бөҖ=0, HвӮӮSбөҖ=0, Pб¶ңбө’бөүб¶ б¶ 
 )

Calculate the lower and upper bounds of the "non-water-selfionization" contributions to total alkalinity.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.Fбө–бҙҙбөӨвӮҷбөўбөЈвӮ’ вҖ” Method

Fбө–бҙҙбөӨвӮҷбөўбөЈвӮ’(AбөҖ, CбөҖ, PбөҖ, SiбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the pH of seawater given the total alkalinity AбөҖ, total carbon CбөҖ, total phosphate PбөҖ, total silicate SiбөҖ, and the carbon chemistry coefficients. Uses the SolveSAPHE package (Munhoven et al., 2013), a universal, robust, pH solver that converges from any given initial value.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮBбөҖ вҖ” Method

function FвҲӮAвҲӮBбөҖ(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮCбөҖ вҖ” Method

FвҲӮAвҲӮCбөҖ(CбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮFбөҖ вҖ” Method

function FвҲӮAвҲӮFбөҖ(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮHвӮӮO вҖ” Method

function FвҲӮAвҲӮHвӮӮO(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮHвӮӮSбөҖ вҖ” Method

function FвҲӮAвҲӮHвӮӮSбөҖ(HвӮӮSбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮNHвӮ„бөҖ вҖ” Method

function FвҲӮAвҲӮNHвӮ„бөҖ(NHвӮ„бөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮPбөҖ вҖ” Method

function FвҲӮAвҲӮPбөҖ(PбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮSOвӮ„бөҖ вҖ” Method

function FвҲӮAвҲӮSOвӮ„бөҖ(HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FвҲӮAвҲӮSiбөҖ вҖ” Method

function FвҲӮAвҲӮSiбөҖ(SiбөҖ, HвҒә, Pб¶ңбө’бөүб¶ б¶ )

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.BOвӮ„HвӮ„вҒ» вҖ” Method

BOвӮ„HвӮ„вҒ»(pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate borate (B(OH)вӮ„вҒ») contribution to Aб¶ң using salinity as a proxy

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.Fбө–бҙҙвҒҪбҙ¬бөҖвҒәбө–б¶ңбө’ВІвҒҫ вҖ” Method

Fбө–бҙҙвҒҪбҙ¬бөҖвҒәбө–б¶ңбө’ВІвҒҫ(AбөҖ, pCOвӮӮ, PбөҖ, SiбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Solve for ocean DIC given total Alkalinity and pCOвӮӮ

Estimate HвҒә (hydrogen ion conc) using estimate of Aб¶ң, carbonate alkalinity after (Follows et al., 2006)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.Fбө–бҙҙвҒҪбҙ¬бөҖвҒәб¶ңбөҖвҒҫ вҖ” Method

Fбө–б¶ңбө’ВІвҒҪбҙ¬бөҖвҒәб¶ңбөҖвҒҫ(AбөҖ, CбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Solve for ocean pCOвӮӮ given total Alkalinity and DIC

Estimate HвҒә (hydrogen ion conc) using estimate of Aб¶ң, carbonate alkalinity after (Follows et al., 2006)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HF вҖ” Method

HF(pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the hydrogen fluoride (HF) contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HPOвӮ„ВІвҒ» вҖ” Method

HPOвӮ„ВІвҒ»(PбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the monohydrogen phosphate (HPOвӮ„ВІвҒ») contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HSOвӮ„вҒ» вҖ” Method

HSOвӮ„вҒ»(pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the hydrogen sulphate (HSOвӮ„вҒ») contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HвҒәб¶ Кібөүбөү вҖ” Method

HвҒәб¶ Кібөүбөү(pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the "Free" HвҒә contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HвӮӮPOвӮ„вҒ» вҖ” Method

HвӮӮPOвӮ„вҒ»(PбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the dihydrogen phosphate (HвӮӮPOвӮ„вҒ») contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HвӮғPOвӮ„ вҖ” Method

HвӮғPOвӮ„(PбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate orthophosphoric acid (HвӮғPOвӮ„) contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.OHвҒ» вҖ” Method

OHвҒ»(pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the hydroxide (OHвҒ») contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.POвӮ„ВівҒ» вҖ” Method

POвӮ„ВівҒ»(PбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the phosphate (POвӮ„ВівҒ») contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.SiOвӮ„HвӮғвҒ» вҖ” Method

SiOвӮ„HвӮғвҒ»(SiбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Calculate the silicate (SiO(OH)вӮғвҒ») contribution to Aб¶ң

ClimaOceanBiogeochemistry.CarbonSystemSolvers.DirectCubicCarbonSolver.Fбө–бҙҙвҒҪбҙ¬бөҖвҒәбө–б¶ңбө’ВІвҒҫ вҖ” Method

Fбө–бҙҙвҒҪбҙ¬бөҖвҒәбө–б¶ңбө’ВІвҒҫ(AбөҖ, pCOвӮӮбөғбө—бөҗ, pH, Pб¶ңбө’бөүб¶ б¶ )

Solve for DIC given total Alkalinity and pCOвӮӮ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.DirectCubicCarbonSolver.Fбө–бҙҙвҒҪбҙ¬бөҖвҒәб¶ңбөҖвҒҫ вҖ” Method

Fбө–бҙҙвҒҪбҙ¬бөҖвҒәб¶ңбөҖвҒҫ(AбөҖ, CбөҖ, pH, Pб¶ңбө’бөүб¶ б¶ )

Solve for ocean pCOвӮӮ given total Alkalinity and DIC