Private types and functions

Documentation for ClimaOceanBiogeochemistry.jl's internal interface.

ClimaOceanBiogeochemistry

ClimaOceanBiogeochemistry.CAN — Method

Tracer sources and sinks for ALK

ClimaOceanBiogeochemistry.CAN — Method

Tracer sources and sinks for DIC

ClimaOceanBiogeochemistry.CAN — Method

Tracer sources and sinks for DOP

ClimaOceanBiogeochemistry.CAN — Method

Tracer sources and sinks for FeT

ClimaOceanBiogeochemistry.CAN — Method

Tracer sources and sinks for NO₃

ClimaOceanBiogeochemistry.CAN — Method

Tracer sources and sinks for PO₄

ClimaOceanBiogeochemistry.CarbonAlkalinityNutrients — Method

CarbonAlkalinityNutrients(; reference_density                           = 1024,
                            maximum_net_community_production_rate       = 1 / day,
                            phosphate_half_saturation                   = 1e-7 * reference_density,
                            nitrate_half_saturation                     = 1.6e-6 * reference_density,
                            iron_half_saturation                        = 1e-10 * reference_density,
                            PAR_half_saturation                         = 10.0,
                            PAR_attenuation_scale                       = 25.0,
                            fraction_of_particulate_export              = 0.33
                            dissolved_organic_phosphate_remin_timescale = 1 / 30day,
                            stoichoimetric_ratio_carbon_to_phosphate    = 106.0
                            stoichoimetric_ratio_nitrate_to_phosphate   = 16.0
                            stoichoimetric_ratio_phosphate_to_oxygen    = 170.0,
                            stoichoimetric_ratio_phosphate_to_iron      = 4.68e-4
                            stoichoimetric_ratio_carbon_to_nitrate      = 106 / 16
                            stoichoimetric_ratio_carbon_to_oxygen       = 106 / 170,
                            stoichoimetric_ratio_carbon_to_iron         = 106 / 1.e-3
                            stoichoimetric_ratio_silicate_to_phosphate  = 15.0
                            rain_ratio_inorganic_to_organic_carbon      = 1e-1
                            martin_curve_exponent                       = 0.84,
                            iron_scavenging_rate                        = 5e-4 / day,
                            ligand_concentration                        = 1e-9 * reference_density,
                            ligand_stability_coefficient                = 1e8)

Return a six-tracer biogeochemistry model for the interaction of carbon, alkalinity, and nutrients.

Keyword Arguments

Tracer names

DIC: Dissolved Inorganic Carbon

Alk: Alkalinity

PO₄: Phosphate (macronutrient)

NO₃: Nitrate (macronutrient)

DOP: Dissolved Organic Phosphate (macronutrient)

Fe: Dissolved Iron (micronutrient)

Biogeochemical functions

transitions for DIC, Alk, PO₄, NO₃, DOP, and Fe

biogeochemical_drift_velocity for D, modeling the sinking of detritus at a constant detritus_sinking_speed.

ClimaOceanBiogeochemistry.NutrientsPlanktonBacteriaDetritus — Method

NutrientsPlanktonBacteriaDetritus(; grid,
                                    maximum_plankton_growth_rate = 1/day,
                                    maximum_bacteria_growth_rate = 1/day
                                    maximum_grazing_rate         = 3/day
                                    bacteria_yield               = 0.2
                                    zooplankton_yield            = 0.3
                                    linear_remineralization_rate = 0.03/day,
                                    linear_mortality_rate        = 0.01/day,
                                    quadratic_mortality_rate     = 0.1/day,
                                    quadratic_mortality_rate_Z   = 1/day,
                                    nutrient_half_saturation     = 0.1,
                                    detritus_half_saturation     = 0.1,
                                    grazing_half_saturation      = 3.0,
                                    PAR_half_saturation          = 10.0,
                                    PAR_attenuation_scale        = 25.0,
                                    detritus_vertical_velocity   = -10/day)

Return a six-tracer biogeochemistry model for the interaction of nutrients (N), phytoplankton (P), zooplankton(Z), bacteria (B), dissolved detritus (D1), and particulate detritus (D2).

Keyword Arguments

grid (required): An Oceananigans' grid.

maximum_plankton_growth_rate: (s⁻¹) Growth rate of plankton P unlimited by the availability of nutrients and light. Default: 1/day.

maximum_bacteria_growth_rate: (s⁻¹) Growth rate of plankton B unlimited by the availability of nutrients and light. Default = 0.5/day.

maximum_grazing_rate: (s⁻¹) Maximum grazing rate of phytoplankton by zooplankton.

bacteria_yield: Determines fractional nutrient production by bacteria production relative to consumption of detritus such that $∂_t N / ∂_t D = 1 - y$, where y = bacteria_yield. Default: 0.2.

linear_remineralization_rate: (s⁻¹) Remineralization rate constant of detritus 'D', assuming linear remineralization of 'D', while implicitly modeling bacteria 'B'. Default = 0.3/day.

linear_mortality_rate: (s⁻¹) Linear term of the mortality rate of both plankton and bacteria.

quadratic_mortality_rate: (s⁻¹) Quadratic term of the mortality rate of both plankton and bacteria.

nutrient_half_saturation: (mmol m⁻³) Half-saturation of nutrients for plankton production.

detritus_half_saturation: (mmol m⁻³) Half-saturation of nutrients for bacteria production. Default = 10.0 mmol m⁻³.

phytoplankton_half_saturation: (mmol m⁻³) Half-saturation of phytoplankton for zooplankton production.

zooplankton_assimilation: Fractional assimilation efficiency for zooplankton.

PAR_half_saturation: (W m⁻²) Half-saturation of photosynthetically available radiation (PAR) for plankton production.

PAR_attenuation_scale: (m) Depth scale over which photosynthetically available radiation (PAR) attenuates exponentially.

detritus_sinking_speed: (m s⁻¹) Sinking velocity of particulate detritus.

Tracer names

N: nutrients

P: phytoplankton

Z: zooplankton

B: bacteria

D1: detritus 1 - dissolved

D2: detritus 2 - particulate

Biogeochemical functions

transitions for N, P, Z, B, D1, D2

biogeochemical_drift_velocity for D2, modeling the sinking of detritus at a constant detritus_sinking_speed.

ClimaOceanBiogeochemistry.iron_scavenging — Method

Iron scavenging should depend on free iron, involves solving a quadratic equation in terms of ligand concentration and stability coefficient, but this is a simple first order approximation.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.CarbonChemistryCoefficients — Method

CarbonChemistryCoefficients(Θᶜ, Sᴬ, Δpᵦₐᵣ)

Return dissociation coefficients necessary to solve for the distribution of carbonate species.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Bᵀᴼᵀ — Function

Bᵀᴼᵀ(Sᵖ, Pᴮᵀᴼᵀ)

Return total borate concentration in mol/kg-SW given practical salinity, Sᵖ. References: Uppström (1974), cited by Dickson et al. (2007, chapter 5, p 10) Millero (1982) cited in Millero (1995)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Caᵀᴼᵀ — Function

Caᵀᴼᵀ(Sᵖ, Pᶜᵃᵀᴼᵀ)

Return calcium concentration in mol/kg-SW given practical salinity, Sᵖ. References: Culkin and Cox (1966), Culkin (1967), Riley and Tongudai (1967)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FCᵀCO₂ˢᵒˡ — Method

FCᵀCO₂ˢᵒˡ(Cᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the dissolved and hydrated CO₂ concentration in seawater given the total carbon concentration Cᵀ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FCᵀCO₃²⁻ — Method

FCᵀCO₃²⁻(Cᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the carbonate concentration in seawater given the total carbon concentration Cᵀ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FCᵀHCO₃⁻ — Method

FCᵀHCO₃⁻(Cᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the bicarbonate ion concentration in seawater given the total carbon concentration Cᵀ, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FpCO₂CO₂ˢᵒˡ — Method

FpCO₂CO₂ˢᵒˡ(pCO₂, pH, Pᶜᵒᵉᶠᶠ)

Calculate the dissolved and hydrated CO₂ concentration in seawater given the pCO₂, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FpCO₂CO₃²⁻ — Method

FpCO₂CO₃²⁻(pCO₂, pH, Pᶜᵒᵉᶠᶠ)

Calculate the carbonate concentration in seawater given the pCO₂, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.FpCO₂HCO₃⁻ — Method

FpCO₂HCO₃⁻(pCO₂, pH, Pᶜᵒᵉᶠᶠ)

Calculate the bicarbonate ion concentration in seawater given the pCO₂, pH, and the carbon chemistry coefficients.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fˢⁱᵗₖ₁ — Function

Fˢⁱᵗₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pˢⁱᵗₖ₁)

Return the first dissociation constant of silicic acid (H4SiO4) in seawater, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pˢⁱᵗₖ₁.

References: Yao and Millero (1995) cited by Millero (1995) pH scale : SWS (according to Dickson et al, 2007) Note : No pressure correction available Note : converted here from mol/kg-H2O to mol/kg-sw

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴴ²ˢₖ₁ — Function

Fᴴ²ˢₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴴ²ˢₖ₁)

Return the dissociation constant of hydrogen sulfide in sea-water, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴴ²ˢₖ₁.

References: Millero et al. (1988) (cited by Millero (1995) Millero (1995) for pressure correction pH scale : - SWS (according to Yao and Millero, 1995, p. 82: "refitted if necessary") - Total (according to Lewis and Wallace, 1998) Note : we stick to SWS here for the time being Note : the fits from Millero (1995) and Yao and Millero (1995) derive from Millero et al. (1988), with all the coefficients multiplied by -ln(10)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴴ²ᴼₖ₁ — Function

Fᴴ²ᴼₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴴ²ᴼₖ₁)

Return dissociation constant of water in (mol/kg-SW)^2, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴴ²ᴼₖ₁.

References: Millero (1995) for value at p_bar = 0 Millero (pers. comm. 1996) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴴˢᴼ⁴ₖ₁ — Function

Fᴴˢᴼ⁴ₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴴˢᴼ⁴ₖ₁)

Return the dissociation constant of hydrogen sulfate (bisulfate) , given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴴˢᴼ⁴ₖ₁.

References: Dickson (1990) – also Handbook (2007) Millero (1995) for pressure correction pH scale : free Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴴᶠᵦ₁ — Function

Fᴴᶠᵦ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴴᶠᵦ₁)

Return the association constant of HF in (mol/kg-SW)^-1, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴴᶠᵦ₁.

HF <-> H⁺ + F⁻

References: Dickson and Riley (1979) Millero (1995) for pressure correction pH scale : free Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴴᶠₖ₁ — Function

Fᴴᶠₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴴᶠₖ₁)

Return the dissociation constant for hydrogen fluoride in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴴᶠₖ₁.

HF <-> H⁺ + F⁻

References: Perez and Fraga (1987) Millero (1995) for pressure correction pH scale : Total (according to Handbook, 2007

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴺᴴ⁴ₖ₁ — Function

Fᴺᴴ⁴ₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴺᴴ⁴ₖ₁)

Return the dissociation constant of ammonium in sea-water [mol/kg-SW], given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴺᴴ⁴ₖ₁.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴾᴼ⁴ₖ₁ — Function

Fᴾᴼ⁴ₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴾᴼ⁴ₖ₁)

Return the first dissociation constant of phosphoric acid (H3PO4) in seawater, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴾᴼ⁴ₖ₁.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴾᴼ⁴ₖ₂ — Function

Fᴾᴼ⁴ₖ₂(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴾᴼ⁴ₖ₂)

Return the second dissociation constant of phosphoric acid (H3PO4) in seawater, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴾᴼ⁴ₖ₂.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᴾᴼ⁴ₖ₃ — Function

Fᴾᴼ⁴ₖ₃(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴾᴼ⁴ₖ₃)

Return the third dissociation constant of phosphoric acid (H3PO4) in seawater, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴾᴼ⁴ₖ₃.

References: Yao and Millero (1995) Millero (1995) for pressure correction pH scale : SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵀᴼᵀ — Function

Fᵀᴼᵀ(Sᵖ, Pᶠᵀᴼᵀ)

Return total fluoride concentration in mol/kg-SW given practical salinity, Sᵖ. References: Culkin (1965) (???)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵃʳᵃᵍᵒⁿⁱᵗᵉₛₚ — Function

Fᵃʳᵃᵍᵒⁿⁱᵗᵉₛₚ(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᵃʳᵃᵍᵒⁿⁱᵗᵉₛₚ)

Return stoichiometric solubility product, Ω, of aragonite in seawater, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᵃʳᵃᵍᵒⁿⁱᵗᵉₛₚ.

References: Mucci (1983) Millero (1979) for pressure correction pH scale : N/A Units : (mol/kg-SW)^2

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵇₖ₁ — Function

Fᵇₖ₁(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᴮₖ₁)

Return boric acid dissociation constant in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᴮₖ₁.

References: Dickson (1990, eq. 23) – also Handbook (2007, eq. 37) Millero (1979) pressure correction pH scale : total

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵈⁱᶜₖ₀ — Function

Fᵈⁱᶜₖ₀(Θᴷ, Sᵖ, Pᵈⁱᶜₖ₀)

Return hydration constant of CO₂ in (mol/kg-SW)/atm given temperature in K, Θᴷ, practical salinity, Sᵖ, and coefficients, Pᵈⁱᶜₖ₀.

CO₂ + H₂O <-> H₂CO₃

References: Weiss (1979) pH scale : N/A Note : currently no pressure correction

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵈⁱᶜₖ₁ᵣ₉₃ — Function

Fᵈⁱᶜₖ₁ᵣ₉₃(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᵈⁱᶜₖ₁ᵣ₉₃)

Return the first dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᵈⁱᶜₖ₁ᵣ₉₃.

H₂CO₃ <-> HCO₃⁻ + H⁺

References: Roy et al. (1993) – also Handbook (1994) Millero (1979) pressure correction pH scale : Total Valid range: T: 0-45 S: 5-45. Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵈⁱᶜₖ₁ₗ₀₀ — Function

Fᵈⁱᶜₖ₁ₗ₀₀(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᵈⁱᶜₖ₁ₗ₀₀)

Return the first dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᵈⁱᶜₖ₁ₗ₀₀.

H₂CO₃ <-> HCO₃⁻ + H⁺

References: Luecker et al. (2000) – also Handbook (2007) Millero (1979) pressure correction pH scale: Total

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵈⁱᶜₖ₁ₘ₉₅ — Function

Fᵈⁱᶜₖ₁ₘ₉₅(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᵈⁱᶜₖ₁ₘ₉₅)

Return the first dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᵈⁱᶜₖ₁ₘ₉₅.

H₂CO₃ <-> HCO₃⁻ + H⁺

References: Millero (1995, eq 50 – ln K1(COM)) Millero (1982) pressure correction pH scale: SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵈⁱᶜₖ₂ᵣ₉₃ — Function

Fᵈⁱᶜₖ₂ᵣ₉₃(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᵈⁱᶜₖ₂ᵣ₉₃)

Return the second dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᵈⁱᶜₖ₂ᵣ₉₃.

HCO₃⁻ <-> CO₃²⁻ + H⁺

References: Roy et al. (1993) – also Handbook (1994) Millero (1979) pressure correction pH scale : Total Valid range: T: 0-45 S: 5-45. Note : converted here from mol/kg-H2O to mol/kg-SW

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵈⁱᶜₖ₂ₗ₀₀ — Function

Fᵈⁱᶜₖ₂ₗ₀₀(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᵈⁱᶜₖ₂ₗ₀₀)

Return the second dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᵈⁱᶜₖ₂ₗ₀₀.

HCO₃⁻ <-> CO₃²⁻ + H⁺

References: Luecker et al. (2000) – also Handbook (2007) Millero (1979) pressure correction pH scale: Total

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᵈⁱᶜₖ₂ₘ₉₅ — Function

Fᵈⁱᶜₖ₂ₘ₉₅(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᵈⁱᶜₖ₂ₘ₉₅)

Return the second dissociation constant of carbonic acid in mol/kg-SW, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᵈⁱᶜₖ₂ₘ₉₅.

HCO₃⁻ <-> CO₃²⁻ + H⁺

References: Millero (1995, eq 51 – ln K2(COM)) Millero (1979) pressure correction pH scale: SWS

ClimaOceanBiogeochemistry.CarbonSystemSolvers.Fᶜᵃˡᶜⁱᵗᵉₛₚ — Function

Fᶜᵃˡᶜⁱᵗᵉₛₚ(Θᴷ, Sᵖ, Δpᵦₐᵣ, Pᶜᵃˡᶜⁱᵗᵉₛₚ)

Return the stoichiometric solubility product of calcite, Ω, in seawater, given temperature in K, Θᴷ, practical salinity, Sᵖ, applied pressure, Δpᵦₐᵣ, and coefficients, Pᶜᵃˡᶜⁱᵗᵉₛₚ

References: Mucci (1983) Millero (1995) for pressure correction pH scale : N/A Units : (mol/kg-SW)^2

ClimaOceanBiogeochemistry.CarbonSystemSolvers.H⁺ₛoverH⁺₃ — Method

H⁺ₛoverH⁺₃(Θᴷ, Sᵖ, Δpᵦₐᵣ)

Return the ratio HSWS/Hfree as a @inline function of salinity, Sᵖ.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.H⁺ₛoverH⁺ₜ — Method

H⁺ₛoverH⁺ₜ(Θᴷ, Sᵖ, Δpᵦₐᵣ)

Return the ratio HSWS/HTot as a @inline function of salinity, Sᵖ. Reference: Munhoven pH scale: all

ClimaOceanBiogeochemistry.CarbonSystemSolvers.H⁺ₜoverH⁺₃ — Method

H⁺ₜoverH⁺₃(Θᴷ, Sᵖ, Δpᵦₐᵣ)

Return the ratio HTot/Hfree as a @inline function of salinity, Sᵖ. Reference: Munhoven pH scale: N/A

ClimaOceanBiogeochemistry.CarbonSystemSolvers.H₂Oˢʷ — Function

H₂Oˢʷ(Sᵖ, Pᴴ²⁰ˢʷ)

Return the mass of pure water in one kg of seawater of practical salinity, Sᵖ. References: "libthdyct" – derived by Munhoven (1997) from data by Millero (1982) "Handbook (2007)" – Handbook (2007) pH scale: N/A

ClimaOceanBiogeochemistry.CarbonSystemSolvers.SO₄ᵀᴼᵀ — Function

SO₄ᵀᴼᵀ(Sᵖ, Pˢᴼ⁴ᵀᴼᵀ)

Return total sulfate concentration in mol/kg-SW given practical salinity, Sᵖ. References: Morris, A.W. and Riley, J.P. (1966) quoted in Handbook (2007)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.μₛ — Function

μₛ(Sᵖ)

Return ionic strength in mol/kg-SW, for given practical salinity, Sᵖ. References: "libthdyct" – derived by Munhoven (1997) from data by Millero (1982) "Handbook (2007)" – Handbook (2007) pH scale: N/A

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FABᵀ — Method

function FABᵀ(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FACᵀ — Method

function FACᵀ(Cᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAFᵀ — Method

function FAFᵀ(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAH₂O — Method

function FAH₂O(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAH₂Sᵀ — Method

function FAH₂Sᵀ(H₂Sᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FANH₄ᵀ — Method

function FANH₄ᵀ(NH₄ᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAPᵀ — Method

function FAPᵀ(Pᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FASO₄ᵀ — Method

function FASO₄ᵀ(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FASiᵀ — Method

function FASiᵀ(Siᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FAᵀ — Method

function FAᵀ(Cᵀ, Aᵀ, Bᵀ, Pᵀ, Siᵀ,  SO₄ᵀ, Fᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

Evaluate the rational function form of the total alkalinity-pH equation

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FH⁺ᵢₙᵢ — Method

FH⁺ᵢₙᵢ(Aᵀ, Cᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

Calculates the root for the 2nd order approximation of the Cᵀ-Bᵀ-Aᶜ equation for H+ around the local minimum, if it exists.

Returns * 1e-03 if Aᶜ <= 0 * 1e-10 if Aᶜ >= 2Cᵀ + Bᵀ * 1e-07 if 0 < Aᶜ < 2Cᵀ + Bᵀ and the 2nd order approximation does not have a solution

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.FboundsAᵀₙₕ₂ₒ — Method

 FboundsAᵀₙₕ₂ₒ(
    Cᵀ, Pᵀ, Siᵀ, NH₄ᵀ=0, H₂Sᵀ=0, Pᶜᵒᵉᶠᶠ
 )

Calculate the lower and upper bounds of the "non-water-selfionization" contributions to total alkalinity.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.Fᵖᴴᵤₙᵢᵣₒ — Function

Fᵖᴴᵤₙᵢᵣₒ(Aᵀ, Cᵀ, Pᵀ, Siᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the pH of seawater given the total alkalinity Aᵀ, total carbon Cᵀ, total phosphate Pᵀ, total silicate Siᵀ, and the carbon chemistry coefficients. Uses the SolveSAPHE package (Munhoven et al., 2013), a universal, robust, pH solver that converges from any given initial value.

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂Bᵀ — Method

function F∂A∂Bᵀ(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂Cᵀ — Method

F∂A∂Cᵀ(Cᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂Fᵀ — Method

function F∂A∂Fᵀ(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂H₂O — Method

function F∂A∂H₂O(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂H₂Sᵀ — Method

function F∂A∂H₂Sᵀ(H₂Sᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂NH₄ᵀ — Method

function F∂A∂NH₄ᵀ(NH₄ᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂Pᵀ — Method

function F∂A∂Pᵀ(Pᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂SO₄ᵀ — Method

function F∂A∂SO₄ᵀ(H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.F∂A∂Siᵀ — Method

function F∂A∂Siᵀ(Siᵀ, H⁺, Pᶜᵒᵉᶠᶠ)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.BO₄H₄⁻ — Method

BO₄H₄⁻(pH, Pᶜᵒᵉᶠᶠ)

Calculate borate (B(OH)₄⁻) contribution to Aᶜ using salinity as a proxy

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.Fᵖᴴ⁽ᴬᵀ⁺ᵖᶜᵒ²⁾ — Method

Fᵖᴴ⁽ᴬᵀ⁺ᵖᶜᵒ²⁾(Aᵀ, pCO₂, Pᵀ, Siᵀ, pH, Pᶜᵒᵉᶠᶠ)

Solve for ocean DIC given total Alkalinity and pCO₂

Estimate H⁺ (hydrogen ion conc) using estimate of Aᶜ, carbonate alkalinity after (Follows et al., 2006)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.Fᵖᴴ⁽ᴬᵀ⁺ᶜᵀ⁾ — Method

Fᵖᶜᵒ²⁽ᴬᵀ⁺ᶜᵀ⁾(Aᵀ, Cᵀ, pH, Pᶜᵒᵉᶠᶠ)

Solve for ocean pCO₂ given total Alkalinity and DIC

Estimate H⁺ (hydrogen ion conc) using estimate of Aᶜ, carbonate alkalinity after (Follows et al., 2006)

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HF — Method

HF(pH, Pᶜᵒᵉᶠᶠ)

Calculate the hydrogen fluoride (HF) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HPO₄²⁻ — Method

HPO₄²⁻(Pᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the monohydrogen phosphate (HPO₄²⁻) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.HSO₄⁻ — Method

HSO₄⁻(pH, Pᶜᵒᵉᶠᶠ)

Calculate the hydrogen sulphate (HSO₄⁻) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.H⁺ᶠʳᵉᵉ — Method

H⁺ᶠʳᵉᵉ(pH, Pᶜᵒᵉᶠᶠ)

Calculate the "Free" H⁺ contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.H₂PO₄⁻ — Method

H₂PO₄⁻(Pᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the dihydrogen phosphate (H₂PO₄⁻) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.H₃PO₄ — Method

H₃PO₄(Pᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate orthophosphoric acid (H₃PO₄) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.OH⁻ — Method

OH⁻(pH, Pᶜᵒᵉᶠᶠ)

Calculate the hydroxide (OH⁻) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.PO₄³⁻ — Method

PO₄³⁻(Pᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the phosphate (PO₄³⁻) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.SiO₄H₃⁻ — Method

SiO₄H₃⁻(Siᵀ, pH, Pᶜᵒᵉᶠᶠ)

Calculate the silicate (SiO(OH)₃⁻) contribution to Aᶜ

ClimaOceanBiogeochemistry.CarbonSystemSolvers.DirectCubicCarbonSolver.Fᵖᴴ⁽ᴬᵀ⁺ᵖᶜᵒ²⁾ — Method

Fᵖᴴ⁽ᴬᵀ⁺ᵖᶜᵒ²⁾(Aᵀ, pCO₂ᵃᵗᵐ, pH, Pᶜᵒᵉᶠᶠ)

Solve for DIC given total Alkalinity and pCO₂

ClimaOceanBiogeochemistry.CarbonSystemSolvers.DirectCubicCarbonSolver.Fᵖᴴ⁽ᴬᵀ⁺ᶜᵀ⁾ — Method

Fᵖᴴ⁽ᴬᵀ⁺ᶜᵀ⁾(Aᵀ, Cᵀ, pH, Pᶜᵒᵉᶠᶠ)

Solve for ocean pCO₂ given total Alkalinity and DIC