Public Documentation

Documentation for ClimaOceanBiogeochemistry.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

ClimaOceanBiogeochemistry

ClimaOceanBiogeochemistry.CarbonSystemSolvers.UniversalRobustCarbonSolver.UniversalRobustCarbonSystem — Method

UniversalRobustCarbonSystem(
        Θ       :: FT = 25.0,
        Sᴬ      :: FT = 35.0,
        Δpᵦₐᵣ   :: FT = 0.0,
        Cᵀ      :: FT = 2050.0e-6,
        Aᵀ      :: FT = 2350.0e-6,
        Pᵀ      :: FT = 1.0e-6,
        Siᵀ     :: FT = 15.0e-6,
        pH      :: FT = 8.0,
        pCO₂ᵃᵗᵐ :: FT = 280.0e-6,
        )

Uses the Munhoven (2013) SolveSAPHE package to solve the distribution of carbon species

ClimaOceanBiogeochemistry.CarbonSystemSolvers.AlkalinityCorrectionCarbonSolver.AlkalinityCorrectionCarbonSystem — Method

AlkalinityCorrectionCarbonSystem(
        Θ       :: FT = 25.0,
        Sᴬ      :: FT = 35.0,
        Δpᵦₐᵣ   :: FT = 0.0,
        Cᵀ      :: FT = 2050.0e-6,
        Aᵀ      :: FT = 2350.0e-6,
        Pᵀ      :: FT = 1.0e-6,
        Siᵀ     :: FT = 15.0e-6,
        pH      :: FT = 8.0,
        pCO₂ᵃᵗᵐ :: FT = 280.0e-6,
        )

Uses the Follows et al (2006) method to solve the distribution of carbon species

ClimaOceanBiogeochemistry.CarbonSystemSolvers.DirectCubicCarbonSolver.DirectCubicCarbonSystem — Method

DirectCubicCarbonSystem(
        Θ       :: FT = 25.0,
        Sᴬ      :: FT = 35.0,
        Δpᵦₐᵣ   :: FT = 0.0,
        Cᵀ      :: FT = 2050.0e-6,
        Aᵀ      :: FT = 2350.0e-6,
        pH      :: FT = 8.0,
        pCO₂ᵃᵗᵐ :: FT = 280.0e-6,
        )

DirectCubicCarbonSolver solves a cubic equation in terms of [H⁺]; Not for serious use, but as a placeholder and for testing purposes